Wrong response_handlers
